Add alloc_vector, split vector_base from vector
